+++ Runtime/OutlineProfile.cs
@@ -7,13 +7,14 @@ namespace Natteens.Outline
     {
         [SerializeField] private bool _enabled = true;
         [SerializeField] private OutlineSelectionMode _selectionMode = OutlineSelectionMode.Targets;
-        [SerializeField, Range(0, 31)] private int _targetRenderingLayerBit = 31;
+        [SerializeField, HideInInspector, Range(0, 31)] private int _targetRenderingLayerBit = 31;
         [SerializeField] private LayerMask _layerMask;
         [SerializeField, Range(1, 8)] private int _thickness = 2;
         [SerializeField] private OutlineColorMode _colorMode = OutlineColorMode.Fixed;
         [SerializeField] private Color _color = Color.black;
         [SerializeField, Range(0f, 1f)] private float _opacity = 1f;
         [SerializeField, Range(0f, 1f)] private float _adaptiveDarken = 0.55f;
+        [SerializeField] private bool _outlineOcclusionEdges;
         [SerializeField] private bool _renderGameCameras = true;
         [SerializeField] private bool _renderSceneView = true;
         [SerializeField] private bool _renderOverlayCameras;
@@ -36,6 +37,7 @@ namespace Natteens.Outline
         public Color Color { get => _color; set => _color = value; }
         public float Opacity { get => _opacity; set => _opacity = Mathf.Clamp01(value); }
         public float AdaptiveDarken { get => _adaptiveDarken; set => _adaptiveDarken = Mathf.Clamp01(value); }
+        public bool OutlineOcclusionEdges { get => _outlineOcclusionEdges; set => _outlineOcclusionEdges = value; }
         public bool RenderGameCameras { get => _renderGameCameras; set => _renderGameCameras = value; }
         public bool RenderSceneView { get => _renderSceneView; set => _renderSceneView = value; }
         public bool RenderOverlayCameras { get => _renderOverlayCameras; set => _renderOverlayCameras = value; }